Restructuring blog posts and add the new IS blog post
