+++ model_app/frontend/src/const/devblogposts.js
@@ -24,14 +24,14 @@ const articles = [
     //url: "https://covidweb.isi.jhu.edu/developmentblog/Post1",
   },
   {
-    id: 5,
-    type: "Announcement",
-    date: "Date 1",
-    author: "Author",
-    title: "React Router",
-    subtext: "Blog 1 Description",
-    href: "/developmentblog/date-3",
-    img: "https://images.pexels.com/photos/325111/pexels-photo-325111.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=750&w=1260",
+    id: 3,
+    type: "Article",
+    date: "July 2021",
+    author: "Serena Chan",
+    title: "Wells-Riley Equation",
+    subtext: "The Delineo simulation uses the Wells-Riley Equation to simulate how a disease will spread in any given city. The equation and the simulation focus on places of interest (POI) that include grocery stores, places of worship, and other types of gathering areas where a person might get infected. We apply...",
+    href: "/developmentblog/Post3",
+    img: "https://i.postimg.cc/Y97qZcxc/1.jpg",
   },
   {
     id: 8,
