+++ model_app/frontend/src/containers/index.js
@@ -3,10 +3,12 @@ import DevelopmentBlog from './DevelopmentBlog';
 import Home from './Home';
 //import Anytown from './Simulator';
 import Team from './Team';
-import Post1, {Post2} from './BlogPost';
+import Post1 from './BlogPost1';
+import Post2 from './BlogPost2';
+import Post3 from './BlogPost3'
 import Simulator from './Simulator';
 //import RealSimulator from './RealSimulator';
 
 
 
-export { About, DevelopmentBlog, Home, Simulator, Team, Post1, Post2}
+export { About, DevelopmentBlog, Home, Simulator, Team, Post1, Post2, Post3}
